# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: gin-template API
  version: 0.1.0
servers:
  - url: /api/v1
components:
  responses:
    ProblemResponse:
      description: Problem
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponseBase:
      type: object
      required: [status, msg]
      properties:
        status: { type: integer, description: 0 表示成功；100000-199999 为系统级状态码；200000-299999 为业务级状态码，按 biz 模块分段 }
        msg: { type: string }
    Problem:
      type: object
      required: [status, msg, details, data]
      properties:
        status: { type: integer }
        msg: { type: string }
        details: { type: string }
        data:
          nullable: true
    User:
      type: object
      required: [uid, username, email, display_name, role, status, email_verified, ctime, utime]
      properties:
        uid: { type: integer, format: int64 }
        username: { type: string }
        email: { type: string }
        display_name: { type: string }
        role: { type: string }
        status: { type: string }
        email_verified: { type: boolean }
        ctime: { type: integer, format: int64, description: Unix 秒级时间戳 }
        utime: { type: integer, format: int64, description: Unix 秒级时间戳 }
...
    UserCreateRequest:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string }
        email: { type: string, format: email }
        password: { type: string }
        display_name: { type: string }
        role: { type: string }
        status: { type: string }
        email_verified: { type: boolean }
    UserUpdateRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string }
        display_name: { type: string }
        role: { type: string }
        status: { type: string }
        email_verified: { type: boolean }
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
        debug_token: { type: string }
    MessageEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/MessageResponse' }
...
    UserListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/User' }
        total: { type: integer }
    UserEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/User' }
    UserListEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/UserListResponse' }
...
paths:
...
  /users/me:
    get:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
    put:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
  /users:
    get:
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserListEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreateRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
  /users/search:
    get:
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserListEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
  /users/{id}:
    get:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
    put:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
    delete:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
...